requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
jinja2==3.1.5
python-multipart==0.0.20
sqlalchemy==2.0.36
alembic==1.14.1
psycopg2-binary==2.9.10
pydantic-settings==2.7.1
httpx==0.28.1
itsdangerous==2.2.0
slowapi==0.1.9
google-api-python-client==2.162.0
google-auth==2.38.0
boto3==1.35.99
bcrypt==5.0.0
cryptography>=42.0
redis==5.2.1
Pillow==11.2.1
apscheduler==3.10.4
